% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lklihd1.R
\name{lklihd1}
\alias{lklihd1}
\title{This is a likelihood function for group 0 (euqal-methylated group)}
\usage{
lklihd1(Ccol, Mcol, alpha1, beta1)
}
\arguments{
\item{Ccol}{a column of read counts for all replicates under two conditions}

\item{Mcol}{a column of methylated read counts for all replicates
under two conditions}

\item{alpha1}{parameter value in likelihood for group 0}

\item{beta1}{parameter value in likelihood for group 0}
}
\value{
the likelihood of data given the site belonging to group 0
}
\description{
This function calculates the likelihood of data given
the site belonging to group 0
}

