% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lklihd2.R
\name{lklihd2}
\alias{lklihd2}
\title{This is a likelihood function for group 1 (hyper-methylated group)}
\usage{
lklihd2(CDcol, CUcol, MDcol, MUcol, integ1, alpha2, beta2)
}
\arguments{
\item{CDcol}{a column of read counts under treatment experiment}

\item{CUcol}{a column of read counts under contrl experiment}

\item{MDcol}{a column of methylated read counts under treatment experiment}

\item{MUcol}{a column of methylated read counts under control experiment}

\item{integ1}{probability of x less than y (x and y follow beta distribtuion
with parameters calculated from the data in treatment and control respectively
please see paper for detail information)}

\item{alpha2}{parameter value for likelihood in group 1}

\item{beta2}{parameter value for likelihood in group 1}
}
\value{
the likelihood of the data given the site belonging to group 1
}
\description{
This function calculates the likelihood of the data
given the site belonging to group 1
}

