% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lklihd3.R
\name{lklihd3}
\alias{lklihd3}
\title{This is a likelihood function for group 1 (hyper-methylated group)}
\usage{
lklihd3(CDcol, CUcol, MDcol, MUcol, integ2, alpha3, beta3)
}
\arguments{
\item{CDcol}{a column of read counts under treatment experiment}

\item{CUcol}{a column of read counts under contrl experiment}

\item{MDcol}{a column of methylated read counts under treatment experiment}

\item{MUcol}{a column of methylated read counts under control experiment}

\item{integ2}{probability of x less than y (x and y follow beta distribtuion
with parameters calculated from the data in treatment and control respectively
please see paper for detail information)}

\item{alpha3}{parameter value for likelihood in group 1}

\item{beta3}{parameter value for likelihood in group 1}
}
\value{
the likelihood of the data given the site belonging to group 1
}
\description{
This function calculates the likelihood of the data
given the site belonging to group 1
}

