% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyBayes.R
\name{MethyBayes}
\alias{MethyBayes}
\title{This is a function to produce methylation identification result
This is a function}
\usage{
MethyBayes(data, n1, n2)
}
\arguments{
\item{data}{a list of read counts , methylated read counts and
sites commonly mapped by all replicates}

\item{n1}{number of replicates in treatment experiment}

\item{n2}{number of replicates in control experiment}
}
\value{
methylation identification result matrix, first
        column is the list of sites, second column is the
        list of identification result
}
\description{
Returns the methylation identification result for all sites
commonly mapped by all replicates under two experimental conditions
}

