% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{A utility function to process the read counts and
methylated read counts to a new data matrix used for MCMC s}
\usage{
process(data, n1, n2)
}
\arguments{
\item{data}{a list of read counts, methylated read counts and commonly mapped sites}

\item{n1}{number of replicates in treatment experiment}

\item{n2}{number of replicates in control experiments}
}
\value{
p matrix used for MCMC
}
\description{
A utility function to process the read counts and
methylated read counts to a new data matrix used for MCMC s
}

