% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdata.R
\name{seqdata}
\alias{seqdata}
\title{This is a function to merge BS-seq data from multiple replicates.}
\usage{
seqdata(cases, controls)
}
\arguments{
\item{cases}{A list of data frames for multiple biological replicates
under treatment experiment.}

\item{controls}{A list of data frames for multiple biological replicates
under control experiment.}
}
\value{
A list of read counts, methylated reads and commonly mapped sites.
}
\description{
Return read counts and methylated reads counts for sites commonly
mapped by all replicates.
}
\examples{
data("case1", "case2", "control1", "control2")
read_counts <- seqdata(list(case1, case2), list(control1, control2))
}

